WIP: setup getCategoryNames function in CollectionService
