+++ src/app/_services/collection.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@angular/core'
 import { HttpClient } from '@angular/common/http'
 
 // Project dependencies
-import { AuthService } from '../shared';
+import { AuthService } from '../shared'
+import { CategoryName } from '../shared/datatypes'
 
 @Injectable({
   providedIn: 'root'
@@ -24,5 +25,21 @@ export class CollectionService {
     return this.http
       .get(this._auth.getUrl() + '/api/v1/categorydesc/' + catId, options)
       .toPromise()
+  }
+
+
+  public getCategoryNames(): Promise<CategoryName[]> {
+    let options = { withCredentials: true }
+
+    return this.http
+      .get(this._auth.getHostname() + '/api/v1/collections/103/categorynames', options)
+      .toPromise()
+      .then(res => {
+        if (res && res[0]) {
+          return <CategoryName[]>res
+        } else {
+          return <CategoryName[]>[]
+        }
+      })
   }
 }
