+++ src/app/browse-page/library.component.ts
@@ -4,12 +4,12 @@ import { Router, ActivatedRoute, Params } from '@angular/router'
 import { Subscription }   from 'rxjs'
 import { map } from 'rxjs/operators'
 
-import { AssetService } from './../shared/assets.service'
-import { AssetSearchService } from './../shared/asset-search.service'
-import { AuthService } from './../shared/auth.service'
+import { AssetService, AssetSearchService, AuthService, TitleService } from './../shared'
+import { CollectionService } from './../_services'
+
 import { TagsService } from './tags.service'
 import { Tag } from './tag/tag.class'
-import { TitleService } from '../shared/title.service'
+
 import { ArtstorStorageService } from '../../../projects/artstor-storage/src/public_api';
 
 @Component({
@@ -69,7 +69,8 @@ export class LibraryComponent implements OnInit {
     private _tags: TagsService,
     private _title: TitleService,
     private _filters: AssetFiltersService,
-    private _storage: ArtstorStorageService
+    private _storage: ArtstorStorageService,
+    private _collectionService: CollectionService
   ) {
     this.unaffiliatedUser = this._auth.isPublicOnly() ? true : false
   }
@@ -153,7 +154,7 @@ export class LibraryComponent implements OnInit {
 
     // Fetch browse collection object from local storage & check if the required collection list has already been set
     let storageBrwseColObj = this._storage.getLocal('browseColObject')
-    
+
     let hasCategoryTitles = storageBrwseColObj && storageBrwseColObj['categoryid'] && storageBrwseColObj['categoryid'][2] && storageBrwseColObj['categoryid'][2].title.length > 0
     if ( storageBrwseColObj && storageBrwseColObj[facetType] && hasCategoryTitles){
       if (facetType === 'artstor-geography'){
@@ -176,7 +177,7 @@ export class LibraryComponent implements OnInit {
         // Categoryid facets require an additional call for labels/titles
         if (facetType == 'categoryid') {
 
-            this._assets.categoryNames()
+            this._collectionService.getCategoryNames()
               .then((data) => {
                 // Create an index by ID for naming the facets
                 let categoryIndex = data.reduce( ( result, item ) => {
