+++ src/app/shared/datatypes/collections/category-name.ts
@@ -1,0 +1,5 @@
+
+export class CategoryName {
+  categoryid: string
+  categoryname: string
+}
