+++ src/app/shared/datatypes/index.ts
@@ -3,3 +3,4 @@ export * from './image-group.interface'
 export * from './thumbnail.interface'
 export * from './type-id.pipe'
 export * from './collection-type-handler'
+export * from './collections/category-name'
